# docs/admin/openapi.yaml
...
openapi: 3.0.3
info:
    title: 用户管理接口
    description: 用户管理
    version: "1.0"
paths:
    /admin/v1/apps:
        get:
            tags:
                - ApplicationService
            description: 获取应用列表
            operationId: ApplicationService_ListApplication
            parameters:
                - name: page
                  in: query
                  description: 当前页码
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  description: 每页的行数
                  schema:
                    type: integer
                    format: int32
                - name: query
                  in: query
                  description: 与过滤参数
                  schema:
                    type: string
                - name: or
                  in: query
                  description: 或过滤参数
                  schema:
                    type: string
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: array
                    items:
                        type: string
                - name: nopaging
                  in: query
                  description: 是否不分页
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListApplicationResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - ApplicationService
            description: 创建应用
            operationId: ApplicationService_CreateApplication
            parameters:
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Application'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Application'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /admin/v1/apps/{id}:
        get:
            tags:
                - ApplicationService
            description: 获取应用数据
            operationId: ApplicationService_GetApplication
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Application'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        put:
            tags:
                - ApplicationService
            description: 更新应用
            operationId: ApplicationService_UpdateApplication
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Application'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Application'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        delete:
            tags:
                - ApplicationService
            description: 删除应用
            operationId: ApplicationService_DeleteApplication
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /admin/v1/login:
        post:
            tags:
                - AuthenticationService
            description: 后台登陆
            operationId: AuthenticationService_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /admin/v1/logout:
        post:
            tags:
                - AuthenticationService
            description: 后台登出
            operationId: AuthenticationService_Logout
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LogoutRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /admin/v1/me:
        get:
            tags:
                - AuthenticationService
            description: 后台获取已经登陆的用户的数据
            operationId: AuthenticationService_GetMe
            parameters:
                - name: id
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
        Application:
            type: object
            properties:
                id:
                    type: integer
                    description: ID
                    format: uint32
                name:
                    type: string
                    description: 应用名称
                appId:
                    type: string
                    description: 应用ID
                appKey:
                    type: string
                    description: 应用密钥
                status:
                    enum:
                        - ON
                        - OFF
                    type: string
                    default: ON
                    description: 应用状态
                creatorId:
                    type: integer
                    description: 创建者ID
                    format: uint32
                ownerId:
                    type: integer
                    description: 拥有者ID
                    format: uint32
                remark:
                    type: string
                    description: 个人描述
                keepMonth:
                    type: integer
                    description: 数据保存多少个月
                    format: uint32
                createTime:
                    type: string
                    description: 创建时间
                updateTime:
                    type: string
                    description: 更新时间
                deleteTime:
                    type: string
                    description: 删除时间
            description: 应用
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        ListApplicationResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Application'
                total:
                    type: integer
                    format: int32
            description: 获取应用列表 - 答复
...
        LoginRequest:
            type: object
            properties:
                userName:
                    type: string
                password:
                    type: string
            description: 用户后台登陆 - 请求
        LoginResponse:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                userName:
                    type: string
                token:
                    type: string
                refreshToken:
                    type: string
            description: 用户后台登陆 - 回应
        LogoutRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
            description: 用户后台登出 - 请求
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        User:
            type: object
            properties:
                id:
                    type: integer
                    description: 用户ID
                    format: uint32
                userName:
                    type: string
                    description: 用户登陆名，使用手机号
                realName:
                    type: string
                    description: 真实姓名
                email:
                    type: string
                    description: 电子邮箱
                avatar:
                    type: string
                    description: 头像
                roleId:
                    type: integer
                    description: 角色ID
                    format: uint32
                creatorId:
                    type: integer
                    description: 创建者ID
                    format: uint32
                authority:
                    enum:
                        - SYS_ADMIN
                        - CUSTOMER_USER
                        - GUEST_USER
                    type: string
                    default: CUSTOMER_USER
                    description: 权限
                description:
                    type: string
                    description: 个人描述
                password:
                    type: string
                    description: 登陆密码
                createTime:
                    type: string
                    description: 创建时间
                updateTime:
                    type: string
                    description: 更新时间
                deleteTime:
                    type: string
                    description: 删除时间
                status:
                    enum:
                        - ON
                        - OFF
                    type: string
                    default: ON
                    description: 用户状态
            description: 用户
...